// lib/schemas.ts
export const getAssignmentHistorySchema = z.object({
  collegeId: z.string().optional(),
  personnelId: z.string().optional(),
  courseId: z.string().optional(),
  scheduleInstanceId: z.string().optional(),
  startDate: z.date().optional(),
  endDate: z.date().optional(),
})